backend/requirements.txt
# AgriFinConnect Rwanda — Backend (Django + ML + Chatbot)
# Django & API
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
drf-yasg>=1.21

# ML (eligibility, risk, loan amount)
joblib>=1.3
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3

# Chatbot: saved T5 model (saved-model/) — Financial_LLM_Chatbot.ipynb
# TFT5ForConditionalGeneration was removed in transformers 5.x; use 4.x for TF T5.
transformers>=4.30,<5.0
# Render may default to newer Python versions where TF wheels are unavailable.
# Keep chatbot optional on unsupported runtimes; app falls back gracefully.
tensorflow>=2.13; python_version < "3.13"
sentencepiece>=0.1.99
huggingface-hub>=0.20

# Translation models (French/Kinyarwanda) — MarianMT via PyTorch
torch>=2.2; python_version < "3.13"

# Production server (Render, etc.)
gunicorn>=21.0
# PostgreSQL on Render (optional; omit for local SQLite)
psycopg2-binary>=2.9
dj-database-url>=2.1
whitenoise>=6.6
